+++ src/Classes/PSModuleGraph.ps1
@@ -1,5 +1,6 @@
 class PSModuleGraph {
     [string]$Path
+    [string]$RootPath
     [object[]]$Nodes
     [object[]]$Edges
     [hashtable]$Analysis